+++ package/src/config/cookie_key.ts
@@ -3,3 +3,5 @@ export const isBotCookieKey = '__is_bot_key__';
 export const isDarkCookieKey = '__is_dark_key__';
 export const cookieConsentCookieKey = '__cookie_consent_key__';
 export const privacyPolicyDateCookieKey = '__privacy_policy_date_key__';
+export const countryCookieKey = '__cf_country__';
+export const timezoneCookieKey = '__cf_timezone__';